Combine policies and views under the same files
