+++ model/src/main/kotlin/io/spine/validation/ValidationPlugin.kt
@@ -63,7 +63,7 @@ public abstract class ValidationPlugin(renderers: List<Renderer<*>> = emptyList(
         SetOnceFieldView::class.java,
     ),
     viewRepositories = setOf<ViewRepository<*, *, *>>(
-        CompiledMessageRepository()
+        CompilationMessageRepository()
     ),
     policies = setOf<Policy<*>>(
         RequiredPolicy(),
