+++ model/src/main/kotlin/io/spine/validation/bound/RangeOption.kt
@@ -27,15 +27,19 @@
 package io.spine.validation.bound
 
 import io.spine.core.External
+import io.spine.core.Subscribe
 import io.spine.core.Where
 import io.spine.option.RangeOption
 import io.spine.protodata.Compilation
+import io.spine.protodata.ast.FieldRef
 import io.spine.protodata.ast.event.FieldOptionDiscovered
 import io.spine.protodata.ast.qualifiedName
 import io.spine.protodata.ast.ref
 import io.spine.protodata.ast.unpack
 import io.spine.protodata.check
 import io.spine.protodata.plugin.Policy
+import io.spine.protodata.plugin.View
+import io.spine.server.entity.alter
 import io.spine.server.event.Just
 import io.spine.server.event.React
 import io.spine.server.event.just
@@ -103,6 +107,22 @@ internal class RangePolicy : Policy<FieldOptionDiscovered>() {
             upperBound = upper.toProto()
             this.file = file
         }.just()
+    }
+}
+
+/**
+ * A view of a field that is marked with the `(range)` option.
+ */
+internal class RangeFieldView : View<FieldRef, RangeField, RangeField.Builder>() {
+
+    @Subscribe
+    fun on(e: RangeFieldDiscovered) = alter {
+        subject = e.subject
+        errorMessage = e.errorMessage
+        range = e.range
+        lowerBound = e.lowerBound
+        upperBound = e.upperBound
+        file = e.file
     }
 }
